Scroll feature added
A snippet of code has been added that scrolls the user to the top of the page after successfully adding an item to their basket. This is especially important for mobile users or for desktop users who are quite far down a product listing page, since it isn't obvious that an item has been added to the basket.